+++ code/core/src/phone/java/com/adobe/marketing/mobile/services/ui/notification/builders/BasicNotificationBuilder.kt
@@ -13,8 +13,10 @@ package com.adobe.marketing.mobile.services.ui.notification.builders
 
 import android.app.Activity
 import android.app.NotificationManager
+import android.app.PendingIntent
 import android.content.BroadcastReceiver
 import android.content.Context
+import android.content.Intent
 import android.view.View
 import android.widget.RemoteViews
 import androidx.core.app.NotificationCompat
@@ -22,7 +24,6 @@ import com.adobe.marketing.mobile.core.R
 import com.adobe.marketing.mobile.services.Log
 import com.adobe.marketing.mobile.services.ServiceProvider
 import com.adobe.marketing.mobile.services.ui.notification.NotificationConstructionFailedException
-import com.adobe.marketing.mobile.services.ui.notification.PendingIntentUtils
 import com.adobe.marketing.mobile.services.ui.notification.PushTemplateConstants
 import com.adobe.marketing.mobile.services.ui.notification.PushTemplateImageUtil
 import com.adobe.marketing.mobile.services.ui.notification.extensions.addActionButtons
@@ -94,7 +95,7 @@ internal object BasicNotificationBuilder {
         notificationBuilder.addActionButtons(
             context,
             trackerActivityClass,
-            pushTemplate.actionButtonsString,
+            pushTemplate.actionButtonsList,
             pushTemplate.tag,
             pushTemplate.isNotificationSticky ?: false
         )
@@ -104,7 +105,7 @@ internal object BasicNotificationBuilder {
             if (pushTemplate.remindLaterEpochTimestamp != null ||
                 pushTemplate.remindLaterDelaySeconds != null
             ) {
-                val remindIntent = PendingIntentUtils.createRemindPendingIntent(
+                val remindIntent = createRemindPendingIntent(
                     context,
                     broadcastReceiverClass,
                     channelIdToUse,
@@ -130,6 +131,133 @@ internal object BasicNotificationBuilder {
             basicPushTemplate,
             trackerActivityClass,
             broadcastReceiverClass
+        )
+    }
+
+    /**
+     * Creates a pending intent for remind later button in a notification.
+     *
+     * @param context the application [Context]
+     * @param broadcastReceiverClass the [Class] of the broadcast receiver to set in the created pending intent
+     * @param channelId [String] containing the notification channel ID
+     * @param pushTemplate the [BasicPushTemplate] object containing the basic push template data
+     * @return the created remind later [PendingIntent]
+     */
+    private fun createRemindPendingIntent(
+        context: Context,
+        broadcastReceiverClass: Class<out BroadcastReceiver>?,
+        channelId: String,
+        pushTemplate: BasicPushTemplate
+    ): PendingIntent? {
+        if (broadcastReceiverClass == null) {
+            return null
+        }
+        Log.trace(
+            PushTemplateConstants.LOG_TAG,
+            SELF_TAG,
+            "Creating a remind later pending intent from a push template object."
+        )
+
+        val remindIntent = Intent(PushTemplateConstants.IntentActions.REMIND_LATER_CLICKED).apply {
+            setClass(context.applicationContext, broadcastReceiverClass)
+
+            flags = Intent.FLAG_ACTIVITY_SINGLE_TOP
+            putExtra(
+                PushTemplateConstants.IntentKeys.TEMPLATE_TYPE, pushTemplate.templateType?.value
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.IMAGE_URI, pushTemplate.imageUrl
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.ACTION_URI, pushTemplate.actionUri
+            )
+            putExtra(PushTemplateConstants.IntentKeys.CHANNEL_ID, channelId)
+            putExtra(
+                PushTemplateConstants.IntentKeys.CUSTOM_SOUND, pushTemplate.sound
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.TITLE_TEXT,
+                pushTemplate.title
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.BODY_TEXT,
+                pushTemplate.body
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.EXPANDED_BODY_TEXT,
+                pushTemplate.expandedBodyText
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.NOTIFICATION_BACKGROUND_COLOR,
+                pushTemplate.notificationBackgroundColor
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.TITLE_TEXT_COLOR,
+                pushTemplate.titleTextColor
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.EXPANDED_BODY_TEXT_COLOR,
+                pushTemplate.expandedBodyTextColor
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.SMALL_ICON, pushTemplate.smallIcon
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.SMALL_ICON_COLOR,
+                pushTemplate.smallIconColor
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.LARGE_ICON, pushTemplate.largeIcon
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.VISIBILITY,
+                pushTemplate.getNotificationVisibility()
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.IMPORTANCE,
+                pushTemplate.getNotificationImportance()
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.BADGE_COUNT, pushTemplate.badgeCount
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.REMIND_EPOCH_TS,
+                pushTemplate.remindLaterEpochTimestamp
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.REMIND_DELAY_SECONDS,
+                pushTemplate.remindLaterDelaySeconds
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.REMIND_LABEL, pushTemplate.remindLaterText
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.ACTION_BUTTONS_STRING,
+                pushTemplate.actionButtonsString
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.STICKY, pushTemplate.isNotificationSticky
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.TAG, pushTemplate.tag
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.TICKER, pushTemplate.ticker
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.PAYLOAD_VERSION, pushTemplate.payloadVersion
+            )
+            putExtra(
+                PushTemplateConstants.IntentKeys.PRIORITY,
+                pushTemplate.notificationPriority
+            )
+        }
+
+        return PendingIntent.getBroadcast(
+            context,
+            0,
+            remindIntent,
+            PendingIntent.FLAG_MUTABLE or PendingIntent.FLAG_UPDATE_CURRENT
         )
     }
 }
