+++ code/core/src/phone/java/com/adobe/marketing/mobile/services/ui/notification/builders/LegacyNotificationBuilder.kt
@@ -71,7 +71,7 @@ internal object LegacyNotificationBuilder {
             .addActionButtons(
                 context,
                 trackerActivityClass,
-                pushTemplate.actionButtonsString,
+                pushTemplate.actionButtonsList,
                 pushTemplate.tag,
                 pushTemplate.isNotificationSticky ?: false
             )
